requirements.txt
Appium-Python-Client
selenium
requests
pandas
pytest
allure-pytest
